requirements.txt
numpy>=1.19.0
scipy>=1.5.0
matplotlib>=3.3.0
pandas>=1.1.0
scikit-learn>=0.23.0
opencv-python>=4.4.0
py-feat>=0.4.0
pyaudio>=0.2.11
webrtcvad>=2.0.10
SpeechRecognition>=3.8.1
pyttsx3>=2.90
python-dotenv>=0.15.0
requests>=2.25.0
openai>=1.0.0
pygame>=2.0.0
rospkg>=1.3.0
cv_bridge
sensor_msgs
